requirements.txt
pandas
scikit-learn
joblib
python-telegram-bot